+++ tests/integration/RequestBuilder/SortingRequestTest.php
@@ -4,7 +4,7 @@ namespace Lmc\Matej\IntegrationTests\RequestBuilder;
 
 use Lmc\Matej\IntegrationTests\IntegrationTestCase;
 use Lmc\Matej\Model\Command\Interaction;
-use Lmc\Matej\Model\Command\Sorting;
+use Lmc\Matej\Model\Command\ItemSorting;
 use Lmc\Matej\Model\Command\UserMerge;
 use Lmc\Matej\Model\Response\SortingResponse;
 
@@ -19,7 +19,7 @@ class SortingRequestTest extends IntegrationTestCase
     {
         $response = static::createMatejInstance()
             ->request()
-            ->sorting(Sorting::create('user-a', ['itemA', 'itemB', 'itemC']))
+            ->sorting(ItemSorting::create('user-a', ['itemA', 'itemB', 'itemC']))
             ->send();
 
         $this->assertInstanceOf(SortingResponse::class, $response);
@@ -32,7 +32,7 @@ class SortingRequestTest extends IntegrationTestCase
     {
         $response = static::createMatejInstance()
             ->request()
-            ->sorting(Sorting::create('user-b', ['item-a', 'item-b', 'itemC-c']))
+            ->sorting(ItemSorting::create('user-b', ['item-a', 'item-b', 'itemC-c']))
             ->setUserMerge(UserMerge::mergeInto('user-b', 'user-a'))
             ->setInteraction(Interaction::withItem('detailview', 'user-a', 'item-a'))
             ->send();
@@ -47,7 +47,7 @@ class SortingRequestTest extends IntegrationTestCase
     {
         $response = static::createMatejInstance()
             ->request()
-            ->sorting(Sorting::create('user-b', ['item-a', 'item-b', 'itemC-c'])->setModelName('invalid-model-name'))
+            ->sorting(ItemSorting::create('user-b', ['item-a', 'item-b', 'itemC-c'])->setModelName('invalid-model-name'))
             ->send();
 
         $this->assertInstanceOf(SortingResponse::class, $response);