+++ tests/unit/RequestBuilder/RequestBuilderFactoryTest.php
@@ -5,7 +5,7 @@ namespace Lmc\Matej\RequestBuilder;
 use Lmc\Matej\Http\RequestManager;
 use Lmc\Matej\Model\Command\ItemProperty;
 use Lmc\Matej\Model\Command\ItemPropertySetup;
-use Lmc\Matej\Model\Command\Sorting;
+use Lmc\Matej\Model\Command\ItemSorting;
 use Lmc\Matej\Model\Command\UserForget;
 use Lmc\Matej\Model\Command\UserItemRecommendation;
 use Lmc\Matej\Model\Request;
@@ -61,7 +61,7 @@ class RequestBuilderFactoryTest extends TestCase
         };
 
         $campaignInit = function (CampaignRequestBuilder $builder): void {
-            $builder->addSorting(Sorting::create('item-id', ['item1', 'item2']));
+            $builder->addSorting(ItemSorting::create('item-id', ['item1', 'item2']));
         };
 
         $forgetInit = function (ForgetRequestBuilder $builder): void {
@@ -93,7 +93,7 @@ class RequestBuilderFactoryTest extends TestCase
                 'sorting',
                 SortingRequestBuilder::class,
                 $voidInit,
-                Sorting::create('user-a', ['item-a', 'item-b', 'item-c']),
+                ItemSorting::create('user-a', ['item-a', 'item-b', 'item-c']),
             ],
             'recommendation' => ['recommendation', RecommendationRequestBuilder::class, $voidInit, $userRecommendation],
             'forget' => ['forget', ForgetRequestBuilder::class, $forgetInit],